requirements.txt
# Plex Bot Dependencies
python-telegram-bot==20.8
python-dotenv==1.0.0
httpx==0.26.0
wakeonlan==3.1.0
APScheduler==3.10.4
paramiko==3.4.0
